add cni-bin-dir parameter to apiserver
This parameter shoudl be added only for Kubernetes >= 1.6 and it breaks
apiserver when added to older version.

The PR is adding kubernetes_version to check it. There may be a problem
with order of salt run checking the version. I'll fix the race
condition in case we will hit it.

PROD-9961

Change-Id: I9c9f42348d4e40e632dcc0a89b535e8aafa88a4a