Enable netchecker to use etcd
Now netchecker is able to use etcd and it is recommended configuration.
TPR became depricated in k8s v.1.7 and will not be used in future versions.
Also, netchecker when using etcd, has more simple and straightforward operation
with persistent storage, with no k8s API in between.
So, it is more stable and resistant. In particular, it is not susceptible to
issues with date variations on different k8s cluster nodes.

Change-Id: Iea1000cb2e79e0fd36dd6959044e13325cc8a163